feat: add basic TOML package parsing
Add bare-bones parsing code for reading and parsing TOML package files. Some code in this commit was mistakenly included in f6c289, as a part of commit splitting.
In addition, add an example package, binutils, for testing. This package was chosen due to it being the first compiled package in LFS.